Updates to allow for SLURM (probably) and SLURM on Cedar

+++ SlurmSubmit.py
@@ -28,7 +28,9 @@ def get_args_with_data_on_node(args, arg_names_to_move, out_dir="$SLURM_TMPDIR")
     s = ""
     args = vars(args)
     for a in arg_names_to_move:
-        if a in args and isinstance(args[a], str) and os.path.exists(args[a]):
+        if (a in args and isinstance(args[a], str)
+            and os.path.exists(args[a])
+            and not "mnist" in args[a] and not "cifar" in args[a]):
             s += f"rsync -rav --relative {args[a]} {out_dir}/\n"
             args[a] = f"{out_dir}/{args[a]}".replace("//", "/").strip("/")
         else:
@@ -50,6 +52,12 @@ def get_slurm_args():
         help="Zero-indexed index of first chunk to run")
     P.add_argument("--last_chunk", default=0, type=int,
         help="Zero-indexed index of last chunk to run, (ie. 9 to run 10 chunks")
+    P.add_argument("--env", default="conda", choices=["conda", "pip"],
+        help="Python environment type")
+    P.add_argument("--gpu_type", default="a100", choices=["a100", "v100l"],
+        help="GPU type")
+    P.add_argument("--mem", default="100G",
+        help="RAM—specify SLURM argument '100G'")
     return P.parse_known_args()
 
 if __name__ == "__main__":
@@ -72,7 +80,12 @@ if __name__ == "__main__":
     else:
         raise NotImplementedError()
 
-    script = f"{file_move_command}\npython {slurm_args.script} {unparse_args(args)} --job_id $SLURM_ARRAY_JOB_ID --num_workers {num_cpus}"
+    script = f"{file_move_command}\npython {slurm_args.script} {unparse_args(args)} --job_id $SLURM_JOB_ID --num_workers {num_cpus}"
+
+    if slurm_args.env == "conda":
+        env_str = "conda activate py3103MRL"
+    elif slurm_args.env == "pip":
+        env_str = "module load python/3.10\nsource ~/virtual_envs/py3103MRL/bin/activate"
 
     with open("slurm/slurm_template.txt", "r") as f:
         slurm_template = f.read()
@@ -81,7 +94,10 @@ if __name__ == "__main__":
     slurm_template = slurm_template.replace("NUM_GPUS", str(num_gpus))
     slurm_template = slurm_template.replace("NUM_CPUS", str(num_cpus))
     slurm_template = slurm_template.replace("NAME", name)
+    slurm_template = slurm_template.replace("PYTHON_ENV_STR", env_str)
     slurm_template = slurm_template.replace("SCRIPT", script)
+    slurm_template = slurm_template.replace("GPU_TYPE", slurm_args.gpu_type)
+    slurm_template = slurm_template.replace("MEM", slurm_args.mem)
 
     slurm_script = f"slurm/{name}.sh"
     with open(slurm_script, "w+") as f:
